Add additional checks to the get_size function to fix Coverity warnings

+++ include/utils.h
@@ -6,4 +6,4 @@ void read_infile(FILE *, unsigned char *, unsigned long long);
 void decrypt_key(unsigned char *, FILE *);
 void read_hdr(struct hdr *, FILE *);
 void write_enc(FILE *, struct hdr *, unsigned char *, char *); 
-unsigned long long get_size(FILE *);
+off_t get_size(FILE *);
